Thêm phân quyền chức năng quản lý thông tin/quản lý tài khoản

+++ VNUA.FITA.MQTT.HRM/VNUA.FITA.MQTT.HRM/Controllers/NhanVienController.cs
@@ -45,14 +45,16 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
         public async Task<IActionResult> Index(string sortOder, string searchString, string currentFilter, int?  pageNumber)
         {
 
-            if (KiemTranChucNang(2) == false & KiemTranChucNang(3)==false)
-            {
-                return RedirectToAction("BaoLoi", "BaoLoi");
-            }
+
 
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
 
             string accconut = HttpContext.Session.GetString("SessionUser");
             ViewData["MaNhanVien"] = String.IsNullOrEmpty(sortOder) ? "id_desc" : "";
@@ -103,6 +105,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -129,6 +137,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             ViewData["IdBP"] = new SelectList(_context.BoPhans, "IdBoPhan", "IdBoPhan");
             return View();
         }
@@ -148,6 +162,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (ModelState.IsValid)
             {
                 if (!KiemTraNhanVien(nhanVien.MaNhanVien))
@@ -179,6 +199,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -204,6 +230,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id != nhanVien.IdNhanVien)
             {
                 return NotFound();
@@ -245,6 +277,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -269,6 +307,12 @@ namespace VNUA.FITA.MQTT.HRM.Controllers
             ViewBag.SessionUser = HttpContext.Session.GetString("SessionUser");
             ViewBag.SessionImage = HttpContext.Session.GetString("SessionImage");
             ViewBag.ChucVu = HttpContext.Session.GetString("SessionChucVu");
+
+            if (KiemTranChucNang(1) == false || KiemTranChucNang(3) == false || KiemTranChucNang(4) == false)
+            {
+                return RedirectToAction("BaoLoi", "BaoLoi");
+            }
+
             var nhanVien = await _context.NhanViens.FindAsync(id);
             _context.NhanViens.Remove(nhanVien);
             await _context.SaveChangesAsync();
